Update MAINTAINERS
Hi all,

Jesse is changing roles internally here at AWS and will no longer be in the GitOps space as frequently as he has been. We've done a handoff and @scottrigby suggested we go ahead with a maintainers file change. To be clear, I've become AWS' point person on GitOps.

Signed-off-by: Chris Short <cbshort@amazon.com>