Link to the documents repo, where the principles are now published
To-do: remove the principles committee in a future update, now that it has folded back into the wider group. My suggestion is to make that change as a separate Git commit, so there's a record.

Remove anchor to the old timeline. GitOps WG timeline moved here (https://github.com/cncf/tag-app-delivery/tree/main/gitops-wg#timeline), but this list on the website is more appropriate to what is described here for the events committee.

Signed-off-by: Scott Rigby <scott@r6by.com>