Update MAINTAINERS
I am offering to volunteer as a maintainer. I have been an active member of the working group. I attend most of the meetings (see meeting notes for attendance verification), was a co-author of the GitOps Principles (and was part of that committee), and I am currently part of the events committee. Being part of the events committee, most of my time there is spent helping out with GitOps Con

Signed-off-by: Christian Hernandez <chernand@redhat.com>